+++ tests/unit/components/cylc/toolbar.vue.spec.js
@@ -18,29 +18,28 @@
 import { mount } from '@vue/test-utils'
 import { createVuetify } from 'vuetify'
 import { createStore } from 'vuex'
+import storeOptions from '@/store/options'
 import Toolbar from '@/components/cylc/Toolbar.vue'
 import WorkflowState from '@/model/WorkflowState.model'
-import storeOptions from '@/store/options'
+
+import CommandMenuPlugin from '@/components/cylc/commandMenu/plugin'
+import sinon from 'sinon'
+import WorkflowService from '@/services/workflow.service'
+import { useDrawer } from '@/utils/toolbar'
+import { vuetifyOptions } from '@/plugins/vuetify'
+import { mdiViewList, mdiBackburger } from '@mdi/js'
+
+const $workflowService = sinon.createStubInstance(WorkflowService)
+const vuetify = createVuetify(vuetifyOptions)
 
 describe('Toolbar component', () => {
-  const vuetify = createVuetify({
-    display: {}
-  })
-  const $route = {
-    name: 'testRoute'
-  }
   let store
-  let mountFunction
+  const { drawer: drawerState } = useDrawer()
 
   beforeEach(() => {
     store = createStore(storeOptions)
-    mountFunction = () => mount(Toolbar, {
-      global: {
-        plugins: [vuetify, store],
-        mocks: { $route }
-      }
-    })
-
+    store.commit('user/SET_USER', { owner: 'rincewind', permissions: ['play', 'pause', 'resume', 'stop'] })
+    drawerState.value = false
     store.state.workflows.workflows = [
       {
         id: 'user/id',
@@ -50,9 +49,83 @@ describe('Toolbar component', () => {
     ]
   })
 
-  it('should mount the component', async () => {
-    const wrapper = mountFunction()
+  it('shows backburger icon when drawer is open and list icon when closed', async () => {
+    const wrapper = mount(Toolbar, {
+      global: {
+        plugins: [store, vuetify, CommandMenuPlugin],
+        mocks: { $workflowService },
+        provide: { versionInfo: null },
+      },
+      props: {
+        views: new Map(),
+        workflowName: 'strewth',
+      },
+    })
+    // Drawer closed -> list icon
+    drawerState.value = false
     await wrapper.vm.$nextTick()
-    expect(wrapper.vm.$el).to.exist
+    const btn = wrapper.find('#toggle-drawer')
+    expect(btn.find('svg path').attributes('d')).to.equal(mdiViewList)
+    // Drawer open -> backburger icon
+    drawerState.value = true
+    await wrapper.vm.$nextTick()
+    expect(btn.find('svg path').attributes('d')).to.equal(mdiBackburger)
+  })
+
+  it('toggles drawer on button click', async () => {
+    const wrapper = mount(Toolbar, {
+      global: {
+        plugins: [store, vuetify, CommandMenuPlugin],
+        mocks: { $workflowService },
+        provide: { versionInfo: null },
+      },
+      props: {
+        views: new Map(),
+        workflowName: 'strewth',
+      },
+    })
+    expect(drawerState.value).to.equal(false)
+    await wrapper.find('#toggle-drawer').trigger('click')
+    expect(drawerState.value).to.equal(true)
+    await wrapper.find('#toggle-drawer').trigger('click')
+    expect(drawerState.value).to.equal(false)
+  })
+
+  it('does not show workflow controls without workflowName', async () => {
+    const wrapper = mount(Toolbar, {
+      global: {
+        plugins: [store, vuetify, CommandMenuPlugin],
+        mocks: { $workflowService },
+        provide: { versionInfo: null },
+      },
+      props: {
+        views: new Map(),
+      },
+    })
+    await wrapper.vm.$nextTick()
+    // Burger button and title should still be present
+    expect(wrapper.find('#toggle-drawer').exists()).to.equal(true)
+    expect(wrapper.find('.c-toolbar-title').exists()).to.equal(true)
+    // Workflow-specific controls should not be present
+    expect(wrapper.find('#workflow-play-button').exists()).to.equal(false)
+    expect(wrapper.find('#workflow-stop-button').exists()).to.equal(false)
+    expect(wrapper.find('#workflow-mutate-button').exists()).to.equal(false)
+    expect(wrapper.find('.c-workflow-controls').exists()).to.equal(false)
+  })
+
+  it('displays the title from the store', async () => {
+    store.commit('app/setTitle', "I'm your pain when you can't feel")
+    const wrapper = mount(Toolbar, {
+      global: {
+        plugins: [store, vuetify, CommandMenuPlugin],
+        mocks: { $workflowService },
+        provide: { versionInfo: null },
+      },
+      props: {
+        views: new Map(),
+      },
+    })
+    await wrapper.vm.$nextTick()
+    expect(wrapper.find('.c-toolbar-title').text()).to.include("I'm your pain when you can't feel")
   })
 })
